requirements.txt
textual>=0.80.0
proxmoxer>=2.0.0
requests>=2.31.0
urllib3>=2.0.0
pyyaml>=6.0
rich>=13.0.0
paramiko>=3.0.0
dnspython>=2.4.0
bcrypt>=4.0.0
